T049_SPD/ml_service/requirements.txt
# Web framework
fastapi==0.118.0
uvicorn==0.37.0
starlette==0.48.0

# Machine Learning and Deep Learning
tensorflow==2.20.0
keras==3.11.3
numpy==2.2.6
tensorboard==2.20.0
tensorboard-data-server==0.7.2

# TensorFlow dependencies
absl-py==2.3.1
astunparse==1.6.3
flatbuffers==25.9.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.1
h5py==3.14.0
libclang==18.1.1
ml-dtypes==0.5.3
namex==0.1.0
opt-einsum==3.4.0
optree==0.17.0
protobuf==6.32.1
termcolor==3.1.0
wrapt==1.17.3

# Image Processing
opencv-python==4.12.0.88
Pillow==11.3.0

# HTTP and File Handling
python-multipart==0.0.20
h11==0.16.0
anyio==4.11.0
sniffio==1.3.1
requests==2.32.5
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10

# Validation and Configuration
pydantic==2.11.9
pydantic-core==2.33.2
annotated-types==0.7.0
typing-extensions==4.15.0
typing-inspection==0.4.2

# Utility libraries
click==8.3.0
six==1.17.0
packaging==25.0
setuptools==80.9.0
wheel==0.45.1

# Rich text and markdown
rich==14.1.0
Markdown==3.9
markdown-it-py==4.0.0
mdurl==0.1.2
MarkupSafe==3.0.3
Pygments==2.19.2

# Web framework additional
Werkzeug==3.1.3

# Note: The following packages from the original requirements.txt are not currently installed:
# scikit-learn, aiofiles, httpx, pandas, structlog, python-dotenv, pydantic-settings
# pytest, pytest-asyncio, pytest-cov, black, flake8, isort
# sqlalchemy, alembic, aiosqlite, python-jose, passlib, slowapi
# marshmallow, python-dateutil, pytz
#
# If you need any of these packages, please install them separately.